Logger moved to OpenProvider, Logger extends AbstractLogger, added default values to Response properties, moved Logger to php-di container
